have updated to double

+++ common.cpp
@@ -105,12 +105,12 @@ int unix_domain_client_init(const char *path){
 
 
 
-void get_send_pos(reader_request_t *reader_request, int *x, int *y){
+void get_send_pos(reader_request_t *reader_request, double *x, double *y){
     *x = reader_request->posx + (reader_request->start_time - reader_request->init_time) * reader_request->velocity;
     *y = reader_request->posy;
 }
 
-void get_recv_pos(reader_request_t *reader_request, tag_response_t *tag_response, int *x, int *y){
+void get_recv_pos(reader_request_t *reader_request, tag_response_t *tag_response, double *x, double *y){
     *x = reader_request->posx + (tag_response->start_time + (tag_response->plen << 3)*1000/UPLINK_BITRATE - reader_request->init_time) * reader_request->velocity;
     *y = reader_request->posy;
 }
